services/ai-engine-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
deepgram-sdk==3.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
aio-pika==9.3.1
httpx==0.26.0
python-dotenv==1.0.1
